108 http methods comment

+++ server/apps/api/core/views.py
@@ -8,6 +8,8 @@ from server.apps.core.utils import *
 from server.apps.api.utils import *
 from rest_framework.decorators import action
 from drf_yasg.utils import swagger_auto_schema
+from server.apps.api.comments.serializers import *
+from server.apps.comments.utils import *
 
 
 class SubmissionViewSet(viewsets.ModelViewSet):
@@ -19,13 +21,15 @@ class SubmissionViewSet(viewsets.ModelViewSet):
     serializer = SubmissionSerializer(queryset, many=True)
     serializer_class = SubmissionSerializer
     filter_backends = (OrderingFilter, SearchFilter)
-    ordering_fields = ['author', 'created_at', 'votes']
+    ordering_fields = ['author', 'created_at', 'votes', 'comments']
     search_fields = ('title', 'url')
     http_method_names = ['get', 'post', 'head', 'delete']
 
     def get_serializer_class(self):
         if self.action == 'create':
             return SubmissionCreateSerializer
+        elif self.action == 'reply':
+            return CommentCreateSerializer
         return SubmissionSerializer
 
     def list(self, request, *args, **kwargs):
@@ -42,7 +46,7 @@ class SubmissionViewSet(viewsets.ModelViewSet):
         response = super(viewsets.ModelViewSet, self).list(request, args, kwargs)
         return response
 
-    @swagger_auto_schema(responses={202: SubmissionSerializer, 404: get_response(ResponseMessages.e404)})
+    @swagger_auto_schema(responses={200: SubmissionSerializer, 404: get_response(ResponseMessages.e404)})
     def retrieve(self, request, *args, **kwargs):
         """
             Shows submission by id
@@ -52,8 +56,9 @@ class SubmissionViewSet(viewsets.ModelViewSet):
         response = super(viewsets.ModelViewSet, self).retrieve(request, args, kwargs)
         return response
 
-    @swagger_auto_schema(responses={202: get_response(desc="Success", message=ResponseMessages.e201_d),
-                                    401: get_response(ResponseMessages.e401), 404: get_response(ResponseMessages.e404)})
+    @swagger_auto_schema(responses={200: get_response(desc="Success", message=ResponseMessages.e201_d),
+                                    401: get_response(ResponseMessages.e401),
+                                    404: get_response(ResponseMessages.e404)})
     def destroy(self, request, *args, **kwargs):
         """
             Deletes a submission
@@ -65,9 +70,13 @@ class SubmissionViewSet(viewsets.ModelViewSet):
         try:
             instance = self.get_object()
             if request.user == instance.author:
-                response_message = {'message': 'Submission has been deleted'}
-                instance.delete()
-                response_status = status.HTTP_202_ACCEPTED
+                if instance.comments == 0:
+                    response_message = {'message': 'Submission has been deleted'}
+                    instance.delete()
+                    response_status = status.HTTP_200_OK
+                else:
+                    response_message = {'message': "ERROR: Can't delete if has comments"}
+                    response_status = status.HTTP_406_NOT_ACCEPTABLE
 
         except Exception as e:
             response_message = {'message': ResponseMessages.e404}
@@ -75,8 +84,10 @@ class SubmissionViewSet(viewsets.ModelViewSet):
 
         return Response(response_message, status=response_status)
 
-    @swagger_auto_schema(responses={202: SubmissionSerializer, 401: get_response(ResponseMessages.e401),
-                                    409: get_response(ResponseMessages.e409), 406: get_response(ResponseMessages.e406)})
+    @swagger_auto_schema(responses={200: SubmissionSerializer,
+                                    401: get_response(ResponseMessages.e401),
+                                    409: get_response(ResponseMessages.e409),
+                                    406: get_response(ResponseMessages.e406)})
     def create(self, request, *args, **kwargs):
         """
             Creates a submission
@@ -109,16 +120,62 @@ class SubmissionViewSet(viewsets.ModelViewSet):
                                           level=0)
                         comment.save()
                         response_message = SubmissionSerializer(submission).data
-                        return Response(response_message, status=status.HTTP_202_ACCEPTED)
+                        return Response(response_message, status=status.HTTP_200_OK)
 
                 submission = standardSaveSubmission(request.user, validated_data)
                 response_message = SubmissionSerializer(submission).data
-                response_status = status.HTTP_202_ACCEPTED
+                response_status = status.HTTP_200_OK
             else:
                 response_message = {'message': ResponseMessages.e406}
                 response_status = status.HTTP_406_NOT_ACCEPTABLE
         return Response(response_message, status=response_status)
 
+    @swagger_auto_schema(responses={200: CommentSerializer, 404: get_response(ResponseMessages.e404)})
+    @action(detail=True, methods=['GET'], name='comments')
+    def comments(self, request, pk, *args, **kwargs):
+        """
+            Shows comments of a submission
+
+            Returns all the comments of a submission
+        """
+        response_status = status.HTTP_404_NOT_FOUND
+        response_message = {'message': ResponseMessages.e404}
+
+        try:
+            submission = Submission.objects.get(id=pk)
+        except Exception as e:
+            return Response(response_message, response_status)
+
+        queryset = Comment.objects.filter(submission=submission)
+        serializer = CommentSerializer(queryset, many=True)
+        return Response(serializer.data)
+
+    @swagger_auto_schema(responses={200: CommentSerializer,
+                                    401: get_response(ResponseMessages.e401),
+                                    404: get_response(ResponseMessages.e404)})
+    @comments.mapping.post
+    def reply(self, request, pk, *args, **kwargs):
+        """
+           Creates a reply of a comment
+
+           Creates a new reply of a comment
+        """
+        response_status = status.HTTP_401_UNAUTHORIZED
+        response_message = {'message': ResponseMessages.e401}
+        if request.user.is_authenticated:
+            try:
+                submission = Submission.objects.get(id=pk)
+            except Exception as e:
+                return Response({'message': ResponseMessages.e404}, status.HTTP_404_NOT_FOUND)
+
+            serializer = CommentCreateSerializer(data=request.data)
+            if serializer.is_valid():
+                reply = create_comment(request.user, serializer.validated_data['text'],
+                                       submission, 0, None)
+                response_message = CommentSerializer(reply).data
+                response_status = status.HTTP_200_OK
+
+        return Response(response_message, status=response_status)
 
 class SubmissionVoteViewSet(viewsets.ModelViewSet):
     """
